+++ src/main/java/com/braintreegateway/Environment.java
@@ -28,6 +28,18 @@ public class Environment {
         this.environmentName = environmentName;
     }
 
+    public static Environment parseEnvironment(String environment) {
+        if (environment.equals("development") || environment.equals("integration")) {
+            return DEVELOPMENT;
+        } else if (environment.equals("sandbox")) {
+            return SANDBOX;
+        } else if (environment.equals("production")) {
+            return PRODUCTION;
+        } else {
+            throw new IllegalArgumentException("Unknown environment: " + environment);
+        }
+    }
+
     private static String developmentBaseURL() {
         if (System.getenv().get("GATEWAY_BASE_URL") != null) {
             return System.getenv().get("GATEWAY_BASE_URL");
@@ -46,5 +58,9 @@ public class Environment {
 
     public String getEnvironmentName() {
         return this.environmentName;
+    }
+
+    public String toString() {
+        return getEnvironmentName();
     }
 }