+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -2004,54 +2004,6 @@ public class TransactionIT implements MerchantAccountTestConstants {
     }
 
     @Test
-    public void submitForSettlementWithOrderIdOnUnsupportedProcessor() {
-        TransactionRequest request = new TransactionRequest().
-            merchantAccountId(FAKE_AMEX_DIRECT_MERCHANT_ACCOUNT_ID).
-            amount(new BigDecimal("100.00")).
-            creditCard().
-                number(CreditCardNumber.AmexPayWithPoints.SUCCESS.number).
-                expirationDate("12/2020").
-                done();
-
-        Transaction transaction = gateway.transaction().sale(request).getTarget();
-
-        TransactionRequest submitForSettlementRequest = new TransactionRequest().
-            orderId("1234");
-
-        Result<Transaction> result = gateway.transaction().submitForSettlement(transaction.getId(), submitForSettlementRequest);
-
-        assertFalse(result.isSuccess());
-        assertEquals(ValidationErrorCode.TRANSACTION_PROCESSOR_DOES_NOT_SUPPORT_UPDATING_ORDER_ID,
-                result.getErrors().forObject("transaction").onField("base").get(0).getCode());
-    }
-
-    @Test
-    public void submitForSettlementWithDescriptorsOnUnsupportedProcessor() {
-        TransactionRequest request = new TransactionRequest().
-            merchantAccountId(FAKE_AMEX_DIRECT_MERCHANT_ACCOUNT_ID).
-            amount(new BigDecimal("100.00")).
-            creditCard().
-                number(CreditCardNumber.AmexPayWithPoints.SUCCESS.number).
-                expirationDate("12/2020").
-                done();
-
-        Transaction transaction = gateway.transaction().sale(request).getTarget();
-
-        TransactionRequest submitForSettlementRequest = new TransactionRequest().
-            descriptor().
-                name("123*123456789012345678").
-                phone("3334445555").
-                url("ebay.com").
-                done();
-
-        Result<Transaction> result = gateway.transaction().submitForSettlement(transaction.getId(), submitForSettlementRequest);
-
-        assertFalse(result.isSuccess());
-        assertEquals(ValidationErrorCode.TRANSACTION_PROCESSOR_DOES_NOT_SUPPORT_UPDATING_DESCRIPTOR,
-                result.getErrors().forObject("transaction").onField("base").get(0).getCode());
-    }
-
-    @Test
     public void submitForSettlementWithBadStatus() {
         TransactionRequest request = new TransactionRequest().
             amount(TransactionAmount.AUTHORIZE.amount).
@@ -4297,6 +4249,57 @@ public class TransactionIT implements MerchantAccountTestConstants {
 
         Transaction refreshedAuthorizedTransaction = gateway.transaction().find(authorizedTransaction.getId());
         assertEquals(2, refreshedAuthorizedTransaction.getPartialSettlementTransactionIds().size());
+    }
+
+    @Test
+    public void submitForPartialSettlementWithOrderId() {
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            creditCard().
+                number(CreditCardNumber.VISA.number).
+                expirationDate("05/2008").
+                done();
+        Transaction transaction = gateway.transaction().sale(request).getTarget();
+
+        BigDecimal amount = new BigDecimal("400.00");
+        TransactionRequest submitForPartialSettlementRequest = new TransactionRequest().
+            amount(amount).
+            orderId("1234");
+
+        Result<Transaction> result = gateway.transaction().submitForPartialSettlement(transaction.getId(), submitForPartialSettlementRequest);
+
+        assertTrue(result.isSuccess());
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, result.getTarget().getStatus());
+        assertEquals("1234", result.getTarget().getOrderId());
+    }
+
+    @Test
+    public void submitForPartialSettlementWithDescriptors()
+    {
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            creditCard().
+                number(CreditCardNumber.VISA.number).
+                expirationDate("05/2008").
+                done();
+        Transaction authorizedTransaction = gateway.transaction().sale(request).getTarget();
+
+        BigDecimal amount = new BigDecimal("400.00");
+        TransactionRequest submitForPartialSettlementRequest = new TransactionRequest().
+            amount(amount).
+            descriptor().
+                name("123*123456789012345678").
+                phone("3334445555").
+                url("ebay.com").
+                done();
+
+        Result<Transaction> result = gateway.transaction().submitForPartialSettlement(authorizedTransaction.getId(), submitForPartialSettlementRequest);
+
+        assertTrue(result.isSuccess());
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, result.getTarget().getStatus());
+        assertEquals("123*123456789012345678", result.getTarget().getDescriptor().getName());
+        assertEquals("3334445555", result.getTarget().getDescriptor().getPhone());
+        assertEquals("ebay.com", result.getTarget().getDescriptor().getUrl());
     }
 
     @Test